Adds build workflow and .gitignore
A build workflow has been added which builds the frontend on each push. Additionally, a .gitignore file has been configured including standard rules for various environments such as Visual Studio, macOS, and Windows. A package.json file has also been created with basic configurations and dependencies for the project, as well as a lockfile for npm packages and a README with install and build instructions.